Added tests RenameFile_should_rename_files, RenameFile_should_rename_directories, RenameFile_should_move_files, RenameFile_should_move_directories and ExtendedLengthFileNameOperations_should_permit_RenameFile_with_a_long_input_path to Interrupt0x21Tests.cs. Fixed RenameFile in DOS.cs to be able to rename directories as well. Annoyingly, the underlying syscall is identical but .NET has File.Move that only works on files and Directory.Move that only works on directories. Corrected the P/Invoke entrypoint in ShortFileNames.Windows.cs for setting a preferred short filename from SetShortPathNameW to SetFileShortNameW.

+++ QBX/OperatingSystem/DOS.cs
@@ -1596,7 +1596,10 @@ public partial class DOS
 				_lastError = DOSError.FileExists;
 			else
 			{
-				File.Move(oldPath, newPath);
+				if (Directory.Exists(oldPath))
+					Directory.Move(oldPath, newPath);
+				else
+					File.Move(oldPath, newPath);
 
 				ShortFileNames.Forget(oldPath);
